requirements.txt
accelerate==0.32.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
auth0-python==4.8.1
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
crewai==0.102.0
cryptography==44.0.2
cycler==0.12.1
decorator==5.2.1
Deprecated==1.2.18
diffusers==0.32.2
distro==1.9.0
docstring_parser==0.16
durationpy==0.9
einops==0.8.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
google-auth==2.38.0
googleapis-common-protos==1.69.0
gradio==3.41.2
gradio_client==0.5.0
groundingdino-py==0.4.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
instructor==1.7.2
ipython==8.33.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
json5==0.10.0
json_repair==0.39.1
jsonpickle==4.0.2
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kubernetes==32.0.1
lazy_loader==0.4
lightning-utilities==0.12.0
litellm==1.60.2
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
narwhals==1.28.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.18.1
openai==1.65.3
opencv-contrib-python-headless==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
posthog==3.18.1
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pygit2==1.15.1
Pygments==2.19.1
PyJWT==2.10.1
PyMatting==1.1.13
pyparsing==3.2.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytorch-lightning==2.3.3
pytz==2025.1
pyvis==0.3.2
PyYAML==6.0.1
referencing==0.36.2
regex==2024.11.6
rembg==2.0.57
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.25.2
scipy==1.14.0
seaborn==0.13.2
segment-anything==1.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.45.3
supervision==0.6.0
sympy==1.13.1
tenacity==9.0.0
tifffile==2025.2.18
tiktoken==0.9.0
timm==1.0.7
tokenizers==0.19.1
tomli==2.2.1
tomli_w==1.2.0
torch==2.1.0
torchmetrics==1.6.1
torchsde==0.2.6
torchvision==0.16.0
tqdm==4.66.4
traitlets==5.14.3
trampoline==0.1.2
transformers==4.42.4
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
update-checker==0.18.0
urllib3==2.3.0
uv==0.6.4
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.17.2
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
